[feature] add IDamageable interface

+++ Assets/Scripts/Units/UnitSO.cs
@@ -7,7 +7,7 @@ public class UnitSO : ScriptableObject
     [field: SerializeField] public string UnitName { get; private set; }
     [field: SerializeField] public GameObject Prefab { get; private set; }
     [field: SerializeField] public float BuildTime { get; private set; }
-    [field: SerializeField] public float Health { get; private set; } = 100f;
+    [field: SerializeField] public int Health { get; private set; } = 100;
     [field: SerializeField] public Sprite Icon { get; private set; }
     [field: SerializeField] public UnitResourceCostSO ResourceCostSO { get; private set; }
 }
